# docs/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "OpenSeat API",
    "version": "0.0.0",
    "description": "フードコートの座席順番待ち。アプリ不要・アカウント不要で、入力は人数だけ。",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
...
  "paths": {
    "/api/v/{venue}/tickets": {
      "post": {
        "operationId": "join",
        "summary": "順番待ちに登録する",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "venue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9](?:[a-z0-9-]{0,30}[a-z0-9])?$"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64
            }
          },
          {
            "name": "x-openseat-client",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "partySize": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 99
                  },
                  "requiredTags": {
                    "default": [],
                    "maxItems": 8,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 32
                    }
                  },
                  "secret": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 128
                  }
                },
                "required": [
                  "partySize",
                  "secret"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "順番待ちに登録する",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "serverNow": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "ticket": {
                      "$ref": "#/components/schemas/TicketView"
                    }
                  },
                  "required": [
                    "serverNow",
                    "ticket"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST / PARTY_SIZE_INVALID / PARTY_TOO_LARGE / PARTY_TOO_SMALL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "JOIN_CLOSED / NO_CODE_AVAILABLE / QUEUE_FULL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{ticket}": {
      "get": {
        "operationId": "readTicket",
        "summary": "チケットの状態を見る",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "ticket",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "チケットの状態を見る",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "serverNow": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "ticket": {
                      "$ref": "#/components/schemas/TicketView"
                    }
                  },
                  "required": [
                    "serverNow",
                    "ticket"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "TICKET_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/t/{ticket}/actions": {
      "post": {
        "operationId": "ticketAction",
        "summary": "本人の操作を送る",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "ticket",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 128
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "cancel"
                      },
                      "reason": {
                        "default": null,
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "found_seat",
                              "leaving",
                              "too_long",
                              "other"
                            ]
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "pause"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "ready"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "extend"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "pass"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "check_in"
                      },
                      "tableToken": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9_-]{16,64}$"
                      }
                    },
                    "required": [
                      "action",
                      "tableToken"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "check_in_early"
                      },
                      "tableToken": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9_-]{16,64}$"
                      }
                    },
                    "required": [
                      "action",
                      "tableToken"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "swap_table"
                      },
                      "tableToken": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9_-]{16,64}$"
                      }
                    },
                    "required": [
                      "action",
                      "tableToken"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "check_out"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "still_here"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "report_conflict"
                      },
                      "tableToken": {
                        "type": "string",
                        "pattern": "^[A-Za-z0-9_-]{16,64}$"
                      }
                    },
                    "required": [
                      "action",
                      "tableToken"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "change_party_size"
                      },
                      "partySize": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 99
                      }
                    },
                    "required": [
                      "action",
                      "partySize"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "const": "heartbeat"
                      }
                    },
                    "required": [
                      "action"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "本人の操作を送る",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "serverNow": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "ticket": {
                      "$ref": "#/components/schemas/TicketView"
                    }
                  },
                  "required": [
                    "serverNow",
                    "ticket"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST / PARTY_SIZE_INVALID / PARTY_TOO_LARGE / PARTY_TOO_SMALL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "STAFF_ONLY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "TABLE_NOT_FOUND / TICKET_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "BLOCKED_BY_GUARD / NOT_ALLOWED_IN_STATE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "TICKET_NOT_FOUND",
              "TABLE_NOT_FOUND",
              "TICKET_ALREADY_EXISTS",
              "PARTY_SIZE_INVALID",
              "PARTY_TOO_SMALL",
              "PARTY_TOO_LARGE",
              "QUEUE_FULL",
              "JOIN_CLOSED",
              "NOT_ALLOWED_IN_STATE",
              "BLOCKED_BY_GUARD",
              "GUARD_NOT_IMPLEMENTED",
              "REASON_REQUIRED",
              "STAFF_ONLY",
              "FORBIDDEN",
              "ACTOR_MISMATCH",
              "NO_CODE_AVAILABLE",
              "CLOCK_WENT_BACKWARD",
              "INVARIANT_VIOLATED",
              "INVALID_REQUEST",
              "UNAUTHORIZED",
              "NOT_FOUND",
              "RATE_LIMITED",
              "INTERNAL"
            ],
            "description": "機械が分岐する識別子"
          },
          "message": {
            "type": "string",
            "description": "利用者に見せる一文。施設のロケールで作られる"
          },
          "retryAfterSec": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "retryAfterSec"
        ],
        "additionalProperties": false
      },
...
      "TicketView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z]-\\d{2}$"
          },
          "state": {
            "type": "string",
            "enum": [
              "WAITING",
              "PAUSED",
              "CALLED",
              "SEATED",
              "DONE",
              "CANCELLED",
              "NO_SHOW",
              "EXPIRED"
            ]
          },
          "partySize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 99
          },
          "requiredTags": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32
            }
          },
          "tableLabel": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ]
          },
          "holdDeadline": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "pauseDeadline": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "extensionsLeft": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "eta": {
            "$ref": "#/components/schemas/WaitEstimate"
          },
          "timeLimit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeLimit"
              },
              {
                "type": "null"
              }
            ]
          },
          "endReason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "checked_out",
                  "staff_checkout",
                  "auto_release",
                  "user_cancel",
                  "staff_cancel",
                  "venue_closed",
                  "no_show",
                  "abandoned",
                  "pause_expired",
                  "max_age"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "JOIN",
                "CANCEL",
                "PAUSE",
                "READY",
                "EXTEND",
                "PASS",
                "CHECK_IN",
                "CHECK_OUT",
                "SWAP_TABLE",
                "CHECK_IN_EARLY",
                "WALK_IN",
                "REPORT_TAKEN",
                "REPORT_IN_USE",
                "CONFIRM_FREE",
                "STILL_HERE",
                "CHANGE_PARTY_SIZE",
                "HEARTBEAT",
                "OPEN",
                "CLOSE",
                "RELEASE_ALL",
                "DISABLE_TABLE",
                "ENABLE_TABLE"
              ]
            }
          }
        },
        "required": [
          "id",
          "code",
          "state",
          "partySize",
          "requiredTags",
          "tableLabel",
          "holdDeadline",
          "pauseDeadline",
          "extensionsLeft",
          "eta",
          "timeLimit",
          "endReason",
          "actions"
        ],
        "additionalProperties": false
      },
      "TimeLimit": {
        "type": "object",
        "properties": {
          "at": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "minutes": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "reached": {
            "type": "boolean"
          }
        },
        "required": [
          "at",
          "minutes",
          "reached"
        ],
        "additionalProperties": false
      },
      "WaitEstimate": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "no_seat"
              }
            },
            "required": [
              "kind"
            ],
            "additionalProperties": false,
            "description": "その人数が収まる席が 1 つも無い"
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "not_waiting"
              }
            },
            "required": [
              "kind"
            ],
            "additionalProperties": false,
            "description": "並んでいない"
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "estimate"
              },
              "minutes": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991,
                "description": "目安の分数（悲観側に切り上げ）"
              },
              "fromMin": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991,
                "description": "表示する幅の下限"
              },
              "toMin": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991,
                "description": "表示する幅の上限"
              },
              "ahead": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991,
                "description": "前に何組いるか"
              }
            },
            "required": [
              "kind",
              "minutes",
              "fromMin",
              "toMin",
              "ahead"
            ],
            "additionalProperties": false
          }
        ]
      }
    }
  }
}